requirements.txt
absl-py==2.3.1
altair==5.5.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
lark==1.3.1
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
narwhals==2.10.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.0
prometheus_client==0.23.1
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
statsmodels==0.14.5
streamlit==1.51.0
sympy==1.14.0
tdqm==0.0.1
tenacity==9.1.2
tensorboard-data-server==0.7.2
termcolor==3.2.0
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
ucimlrepo==0.0.7
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wrapt==2.0.0
